# src/backend/projects/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/synergies": {
            "get": {
                "description": "Retorna uma lista de todas as sinergias",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "synergies"
                ],
                "summary": "Obtém todas as sinergias",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.Synergy"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Cria uma nova sinergia com base nos dados fornecidos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "synergies"
                ],
                "summary": "Cria uma nova sinergia",
                "parameters": [
                    {
                        "description": "Dados da sinergia",
                        "name": "synergy",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.SynergyPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "synergy created successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/synergies/{id}": {
            "get": {
                "description": "Retorna uma sinergia com base no ID fornecido",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "synergies"
                ],
                "summary": "Obtém uma sinergia por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da sinergia",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.Synergy"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Atualiza uma sinergia existente com base nos dados fornecidos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "synergies"
                ],
                "summary": "Atualiza uma sinergia existente",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da sinergia",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados da sinergia",
                        "name": "synergy",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.SynergyPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "synergy updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Exclui uma sinergia existente com base no ID fornecido",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "synergies"
                ],
                "summary": "Exclui uma sinergia existente",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da sinergia",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "synergy deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.Synergy": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "source_project_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "target_project_id": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "types.SynergyPayload": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "source_project_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "target_project_id": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}